koordinatės pakeistos iš double į int

+++ TankoProjektas/src/com/vitkus/tankoprojektas/Point.java
@@ -1,26 +1,26 @@
 package com.vitkus.tankoprojektas;
 
 public class Point {
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
 
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
-	public Point(double x, double y) {
+	public Point(int x, int y) {
 		this.x=x;
 		this.y=y;
 	}
